Test loading of large rank expressions from files
