+++ model-evaluation/src/test/java/ai/vespa/models/evaluation/ModelTester.java
@@ -73,39 +73,4 @@ public class ModelTester {
         assertEquals(expression, function.getBody().getRoot().toString());
     }
 
-    /** Allows us to provide canned tensor constants during import since file distribution does not work in tests */
-    public static class RankProfilesConfigImporterWithMockedConstants extends RankProfilesConfigImporter {
-
-        private static final Logger log = Logger.getLogger(RankProfilesConfigImporterWithMockedConstants.class.getName());
-
-        private final Path constantsPath;
-
-        public RankProfilesConfigImporterWithMockedConstants(Path constantsPath, FileAcquirer fileAcquirer) {
-            super(fileAcquirer);
-            this.constantsPath = constantsPath;
-        }
-
-        @Override
-        protected Tensor readTensorFromFile(String name, TensorType type, FileReference fileReference) {
-            try {
-                return TypedBinaryFormat.decode(Optional.of(type),
-                                                GrowableByteBuffer.wrap(IOUtils.readFileBytes(constantsPath.append(name).toFile())));
-            }
-            catch (IOException e) {
-                log.warning("Missing a mocked tensor constant for '" + name + "': " + e.getMessage() +
-                            ". Returning an empty tensor");
-                return Tensor.from(type, "{}");
-            }
-        }
-
-        @Override
-        protected RankingExpression readExpressionFromFile(String name, FileReference fileReference) throws ParseException {
-            try {
-                return new RankingExpression(name, Utf8.toString(IOUtils.readFileBytes(constantsPath.append(name).toFile())));
-            } catch (IOException e) {
-                throw new IllegalArgumentException("Missing expression file '" + name + "'", e);
-            }
-        }
-    }
-
 }
