+++ model-evaluation/src/test/resources/config/rankexpression/ranking-expressions.cfg
@@ -1,0 +1,2 @@
+expression[0].name "overflow.firstphase"
+expression[0].fileref "overflow.firstphase.expr"
